feat(screenshot): add macOS ScreenCaptureKit capture support
- capture displays and windows with point-based geometry, cursor support, cancellation, and permission recovery
- preserve mixed-display backing scales across selection, annotation, export, scrolling, and history
- add macOS native smoke tests, rendering coverage, validation docs, and persistent code-signing support
- hide Windows-only screenshot settings on macOS while retaining stored configuration

+++ snow_shot/tests/direct_capture_frame_tests.cpp
@@ -57,7 +57,7 @@ int main() {
                 pixels[pixel * 4 + 1],
                 pixelFormat == SNOW_CAPTURE_PIXEL_FORMAT_BGRA8 ? pixels[pixel * 4]
                                                                : pixels[pixel * 4 + 2],
-                alpha[pixel]);
+                alpha[static_cast<std::size_t>(pixel)]);
             require(wrapped.pixelColor(pixel % 2, pixel / 2) == expected &&
                         converted.pixelColor(pixel % 2, pixel / 2) == expected,
                     "native frame conversion changed colors or transparency");
